# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Hepatica API
  version: 0.1.0
servers:
  - url: http://localhost:8000
paths:
  /api/v1/patients:
    post:
      tags: [patients]
      summary: Create patient profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientCreate'
      responses:
        '201':
          description: Patient created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientRead'
  /api/v1/patients/{patientId}:
    get:
      tags: [patients]
      summary: Fetch patient summary
      parameters:
        - name: patientId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Patient record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientRead'
  /api/v1/assessments/clinical:
    post:
      tags: [assessments]
      summary: Compute Stage 1 clinical risk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClinicalAssessmentCreate'
      responses:
        '200':
          description: Stage 1 result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClinicalAssessmentRead'
  /api/v1/scans/upload-url:
    post:
      tags: [scans]
      summary: Create pre-signed URL for scan upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Upload ticket
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
  /api/v1/assessments/fibrosis:
    post:
      tags: [assessments]
      summary: Run Stage 2 fibrosis inference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FibrosisAssessmentCreate'
      responses:
        '200':
          description: Stage 2 prediction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FibrosisAssessmentRead'
  /api/v1/knowledge/explain:
    post:
      tags: [knowledge]
      summary: Generate structured explanation with citations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeExplainRequest'
      responses:
        '200':
          description: Explanation blocks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeExplainResponse'
  /api/v1/reports:
    post:
      tags: [reports]
      summary: Generate patient report PDF
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportCreate'
      responses:
        '200':
          description: Report object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportRead'
  /api/v1/reports/{reportId}:
    get:
      tags: [reports]
      summary: Fetch report metadata and link
      parameters:
        - name: reportId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Report metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportRead'
  /api/v1/patients/{patientId}/timeline:
    get:
      tags: [timeline]
      summary: Fetch patient timeline
      parameters:
        - name: patientId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Timeline events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineRead'
components:
  schemas:
    RiskTier:
      type: string
      enum: [LOW, MODERATE, HIGH]
    FibrosisStage:
      type: string
      enum: [F0, F1, F2, F3, F4]
    ConfidenceFlag:
      type: string
      enum: [NORMAL, LOW_CONFIDENCE]
    EscalationFlag:
      type: string
      enum: [NONE, SEVERE_STAGE_REVIEW]
    PatientCreate:
      type: object
      required: [external_id]
      properties:
        external_id: { type: string }
        sex: { type: string }
        age: { type: integer }
        bmi: { type: number }
        type2dm: { type: boolean }
        notes: { type: string }
    PatientRead:
      allOf:
        - $ref: '#/components/schemas/PatientCreate'
        - type: object
          required: [id, created_at]
          properties:
            id: { type: string }
            created_at: { type: string, format: date-time }
    ClinicalAssessmentCreate:
      type: object
      required: [patient_id, ast, alt, platelets, ast_uln, age, bmi, type2dm]
      properties:
        patient_id: { type: string }
        ast: { type: number }
        alt: { type: number }
        platelets: { type: number }
        ast_uln: { type: number }
        age: { type: integer }
        bmi: { type: number }
        type2dm: { type: boolean }
    ClinicalAssessmentRead:
      type: object
      required: [id, patient_id, fib4, apri, risk_tier, probability, model_version, created_at]
      properties:
        id: { type: string }
        patient_id: { type: string }
        fib4: { type: number }
        apri: { type: number }
        risk_tier: { $ref: '#/components/schemas/RiskTier' }
        probability: { type: number }
        model_version: { type: string }
        created_at: { type: string, format: date-time }
    UploadUrlRequest:
      type: object
      required: [patient_id, filename, content_type, byte_size]
      properties:
        patient_id: { type: string }
        filename: { type: string }
        content_type: { type: string }
        byte_size: { type: integer }
    UploadUrlResponse:
      type: object
      required: [scan_asset_id, object_key, upload_url, expires_in_seconds]
      properties:
        scan_asset_id: { type: string }
        object_key: { type: string }
        upload_url: { type: string }
        expires_in_seconds: { type: integer }
    FibrosisAssessmentCreate:
      type: object
      required: [patient_id, scan_asset_id]
      properties:
        patient_id: { type: string }
        scan_asset_id: { type: string }
    FibrosisAssessmentRead:
      type: object
      required:
        [prediction_id, patient_id, scan_asset_id, model_version, softmax_vector, top1, top2, confidence_flag, escalation_flag, quality_metrics, created_at]
      properties:
        prediction_id: { type: string }
        patient_id: { type: string }
        scan_asset_id: { type: string }
        model_version: { type: string }
        softmax_vector: { type: object, additionalProperties: { type: number } }
        top1:
          $ref: '#/components/schemas/TopPrediction'
        top2:
          type: array
          items:
            $ref: '#/components/schemas/TopPrediction'
        confidence_flag:
          $ref: '#/components/schemas/ConfidenceFlag'
        escalation_flag:
          $ref: '#/components/schemas/EscalationFlag'
        quality_metrics:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
    TopPrediction:
      type: object
      required: [stage, probability]
      properties:
        stage: { $ref: '#/components/schemas/FibrosisStage' }
        probability: { type: number }
    KnowledgeExplainRequest:
      type: object
      required: [patient_id]
      properties:
        patient_id: { type: string }
        fibrosis_stage: { $ref: '#/components/schemas/FibrosisStage' }
        top_k: { type: integer }
    Citation:
      type: object
      required: [source_doc, page_number]
      properties:
        source_doc: { type: string }
        page_number: { type: integer }
    KnowledgeBlock:
      type: object
      required: [title, content, citations]
      properties:
        title: { type: string }
        content: { type: string }
        citations:
          type: array
          items:
            $ref: '#/components/schemas/Citation'
    KnowledgeExplainResponse:
      type: object
      required: [patient_id, blocks]
      properties:
        patient_id: { type: string }
        blocks:
          type: array
          items:
            $ref: '#/components/schemas/KnowledgeBlock'
    ReportCreate:
      type: object
      required: [patient_id]
      properties:
        patient_id: { type: string }
        clinical_assessment_id: { type: string }
        fibrosis_prediction_id: { type: string }
    ReportRead:
      type: object
      required: [report_id, patient_id, report_json, created_at]
      properties:
        report_id: { type: string }
        patient_id: { type: string }
        pdf_download_url: { type: string }
        report_json: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
    TimelineEventRead:
      type: object
      required: [id, patient_id, event_type, event_payload, created_at]
      properties:
        id: { type: string }
        patient_id: { type: string }
        event_type: { type: string }
        event_payload: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
    TimelineRead:
      type: object
      required: [patient_id, events]
      properties:
        patient_id: { type: string }
        events:
          type: array
          items:
            $ref: '#/components/schemas/TimelineEventRead'